add register control

+++ src/main/java/com/mutopia/sys/model/user/SysUserRoleRel.java
@@ -18,7 +18,7 @@ public class SysUserRoleRel extends com.mutopia.sys.model.base.BaseEntity  {
 	private Date createTime;
 
 	@Column(name="create_user")
-	private int createUser;
+	private long createUser;
 
 	@Column(name="role_id")
 	private int roleId;
@@ -26,7 +26,7 @@ public class SysUserRoleRel extends com.mutopia.sys.model.base.BaseEntity  {
 	private String status;
 
 	@Column(name="user_id")
-	private int userId;
+	private long userId;
 
 	public SysUserRoleRel() {
 	}
@@ -39,11 +39,11 @@ public class SysUserRoleRel extends com.mutopia.sys.model.base.BaseEntity  {
 		this.createTime = createTime;
 	}
 
-	public int getCreateUser() {
+	public long getCreateUser() {
 		return this.createUser;
 	}
 
-	public void setCreateUser(int createUser) {
+	public void setCreateUser(long createUser) {
 		this.createUser = createUser;
 	}
 
@@ -63,11 +63,11 @@ public class SysUserRoleRel extends com.mutopia.sys.model.base.BaseEntity  {
 		this.status = status;
 	}
 
-	public int getUserId() {
+	public long getUserId() {
 		return this.userId;
 	}
 
-	public void setUserId(int userId) {
+	public void setUserId(long userId) {
 		this.userId = userId;
 	}
